+++ nix/package/build.sh
@@ -1,0 +1,3 @@
+# shellcheck shell=bash
+set -euo pipefail
+typst compile --input "rev=$REV" cv.typ cv.pdf